requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
prometheus-client==0.19.0
pydantic==2.5.2
pydantic-settings==2.1.0

